Do not use custom elements in the popup.
