+++ src/popup/popup-login.js
@@ -1,13 +1,10 @@
-import {customElement, noShadowDOM} from 'solid-element'
 import {createSignal, createResource, Show} from 'solid-js'
 
-import Config from '../config.js'
+import config from '../config.js'
 import store from '../store/store-content.js'
-import './popup-login-form.js'
+import PopupLoginForm from './popup-login-form.js'
 
-customElement('popup-login', {}, () => {
-  noShadowDOM()
-
+export default function PopupLogin () {
   const [getSession, setGetSession] = createSignal(false)
   const [showLoginForm] = createResource(getSession, () => {
     // check session
@@ -25,13 +22,13 @@ customElement('popup-login', {}, () => {
   return (
     <div class="popup-login text-center">
       <div class="popup-box popup-logo">
-        <a href={Config.websiteUrl} target="_blank">
+        <a href={config.websiteUrl} target="_blank">
           <img src="../icons/briskine-combo.svg" width="160" alt="Briskine"/>
         </a>
       </div>
 
       <div class="popup-box">
-        <Show when={!showLoginForm()} fallback={(<popup-login-form />)}>
+        <Show when={!showLoginForm()} fallback={(<PopupLoginForm />)}>
           <p>
             <strong>
               Sign in to access your templates.
@@ -54,11 +51,11 @@ customElement('popup-login', {}, () => {
         <small>
           Don't have an account yet?
           <br />
-          <a href={`${Config.functionsUrl}/signup/`} target="_blank">
+          <a href={`${config.functionsUrl}/signup/`} target="_blank">
             Create a free account
           </a>
         </small>
       </div>
     </div>
   )
-})
+}
