+++ src/popup/popup.html
@@ -7,8 +7,7 @@
     <link href="/popup/popup.css" rel="stylesheet">
   </head>
   <body>
-    <popup-container></popup-container>
-
+    <div id="app"></div>
     <script src="popup.js"></script>
   </body>
 </html>
